Charbel's change
change the color of the main gradient theme, added a glass effect on hover. Added a sub menu when clicking on the profile. A green border appears around the profile to indicate that the person is in the office. When he's on leave, the border is red. I still need to work on the animations and the colors of different buttons in the portal